Joe/lang 194 implement model specific prompt formatting for cog triton

+++ utils.py
@@ -13,6 +13,11 @@ import subprocess
 from pathlib import Path
 import logging
 
+import argparse
+import subprocess
+import sys
+from pathlib import Path
+
 
 def maybe_download_tarball_with_pget(
     url: str,
@@ -30,31 +35,21 @@ def maybe_download_tarball_with_pget(
         path (str): Path to the directory where files were downloaded
 
     """
-    print("Checking for tarball...")
-    # if dest exists, rm
+
+    # if dest exists and is not empty, return
+    if os.path.exists(dest) and os.listdir(dest):
+        print(f"Files already present in the `{dest}`, nothing will be downloaded.")
+        return dest
+
+    # if dest exists but is empty, remove it so we can pull with pget
     if os.path.exists(dest):
         shutil.rmtree(dest)
 
-    print("Downloading weights...")
+    print("Downloading model assets...")
     command = ["pget", url, dest, "-x"]
     subprocess.check_call(command, close_fds=True)
 
     return dest
-
-    # subprocess.run(
-    #     [
-    #         "python3",
-    #         "/src/tensorrtllm_backend/scripts/launch_triton_server.py",
-    #         "--world_size=1",
-    #         "--model_repo=/src/triton_model_repo",
-    #     ]
-    # )
-
-
-import argparse
-import subprocess
-import sys
-from pathlib import Path
 
 
 class TritonHandler:
